requirements.txt
# requirements.txt - Sistema Multi-Agente RAG Local
# Python 3.10+

# LLM inference
openai>=1.35.0

# LangChain
langchain>=0.2.14
langchain-openai>=0.1.22
langchain-community>=0.2.14
langchain-text-splitters>=0.2.2

# Embeddings
sentence-transformers>=3.0.1

# Vector store
faiss-cpu>=1.8.0

# PDF processing
pymupdf>=1.24.9

# API
fastapi>=0.112.0
uvicorn[standard]>=0.30.6
pydantic>=2.8.0
pydantic-settings>=2.4.0

# UI
streamlit>=1.37.0
requests>=2.32.0

# Data
numpy>=1.26.4
pandas>=2.2.2

# Utilities
python-dotenv>=1.0.1
tqdm>=4.66.5
httpx>=0.27.0

# Testing
pytest>=8.3.2
pytest-asyncio>=0.23.8

# Notebooks and analysis
jupyter>=1.0.0
matplotlib>=3.9.2
seaborn>=0.13.2
umap-learn>=0.5.6
